fix(campaigns): pipe the aria label and announce search status
The previous commit's a11y fix reintroduced the exact violation the commit two
before it had removed: a template method call, in the same file, over the same
500 rows. frontend-checklist.md §4 allows only signal reads, computed values and
pipes — and emailTemplateAccessibleName did MORE work per row than the call it
replaced, because it invoked formatHubSpotUpdatedAt internally.

Its stated justification was that an Angular template cannot pipe inside a
ternary. True, but that argues for a pipe taking the whole template object, which
sidesteps the ternary entirely. Cursor reached the same remedy independently.

Search status was also silent. A persistent live region now carries it, following
implementation-tab's documented pattern: a region inserted with its text already
present is not reliably announced, so it must exist first and have its CONTENTS
change.

LFXV2-3198
